[AV2-228] fixed issues found in code review.
